updating_tables_daily/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-api-python-client
pandas
pandas-gbq
PyYAML
python-dateutil